// src/shared/types.ts
export type configSchemaType = z.infer<typeof z$configSchemaType>;
// ...
export type configDatabaseType = z.infer<typeof z$configDatabaseType>;

export const z$configHostType = z.object({
	enabled: z.boolean(),
	name: z.string(),
	address: z.string(),
	probe: z.boolean(),
	databases: z.array(z$configDatabaseType),
}).strict();
export type configHostType = z.infer<typeof z$configHostType>;

export const z$configCustomStatType = z.object({
	title: z.string(),
	sql: z.string(),
}).strict();
export type configCustomStatType = z.infer<typeof z$configCustomStatType>;
export const z$configCustomRepository = z.record(z.string().min(1), z.array(z$configCustomStatType));
export type configCustomRepository = z.infer<typeof z$configCustomRepository>;

export const z$configOptionsType = z.object({
	pollingSeconds: z.number(),
	hidePasswords: z.boolean(),
	connectTimeoutSeconds: z.number(),
}).strict();
export type configOptionsType = z.infer<typeof z$configOptionsType>;
// ...
export type flatSchemaType = Omit<configSchemaType, 'customSelect'>;
export type flatDatabaseType = Omit<configDatabaseType, 'customStats' | 'customSelect' | 'schemas'>;
export type flatHostType = Omit<configHostType, 'databases'>;
export type flatType = {
	host: flatHostType,
	database: flatDatabaseType,
	schema?: flatSchemaType,
};